+++ dwds/lib/src/debugging/metadata/provider.dart
@@ -65,7 +65,6 @@ class MetadataProvider {
         'dart:svg',
         'dart:web_audio',
         'dart:web_gl',
-        'dart:web_sql',
         'dart:ui',
       ];